refactor(gradle): Removed unused tests/resource files, relocated necessary tests/resource files
