--- detectable/src/test/resources/detectables/functional/gradle/maven-metadata.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<metadata>
-    <groupId>com.blackducksoftware.integration</groupId>
-    <artifactId>integration-gradle-inspector</artifactId>
-    <versioning>
-        <latest>1.1.0</latest>
-        <release>1.1.0</release>
-        <versions>
-            <version>0.0.1</version>
-            <version>0.0.2</version>
-            <version>0.0.3</version>
-            <version>0.0.4</version>
-            <version>0.0.5</version>
-            <version>0.0.6</version>
-            <version>0.0.7</version>
-            <version>0.1.0</version>
-            <version>0.2.0</version>
-            <version>0.2.1</version>
-            <version>0.2.2</version>
-            <version>0.3.0</version>
-            <version>0.3.1</version>
-            <version>0.4.0</version>
-            <version>0.5.0</version>
-            <version>0.5.1</version>
-            <version>0.6.0</version>
-            <version>0.7.0</version>
-            <version>1.0.0</version>
-            <version>1.1.0</version>
-        </versions>
-        <lastUpdated>20180726195440</lastUpdated>
-    </versioning>
-</metadata>
