--- detectable/src/test/resources/detectables/functional/gradle/reports/Gradle_Report_NestedProjectDependency.txt
@@ -1,9 +1,0 @@
-
-------------------------------------------------------------
-Tests that a project dependency with a child project dependency works
-------------------------------------------------------------
-
-nestedDependency - this configuration contains a project dependency with another project dependency
-+--- project :nested-parent
-|    \--- project :nested-child
-\--- example:normal-component:1.0.0
