--- detectable/src/test/resources/detectables/functional/gradle/reports/Gradle_Report_NoDependencies.txt
@@ -1,7 +1,0 @@
-
-------------------------------------------------------------
-Tests that we handle no dependencies
-------------------------------------------------------------
-
-noDependency - this configuration has no dependencies
-No dependencies
