--- detectable/src/test/resources/detectables/functional/gradle/reports/Gradle_Report_ProjectDependency.txt
@@ -1,7 +1,0 @@
-
-------------------------------------------------------------
-Tests that a project dependency does not show up.
-------------------------------------------------------------
-
-projectTest - this configuration has a single project dependency that should not appear
-\--- project :child-project
