--- detectable/src/test/resources/detectables/functional/gradle/spring-framework/spring_aspects_dependencyGraph.txt
@@ -1,204 +1,0 @@
-
-------------------------------------------------------------
-Project :spring-aspects - Spring Aspects
-------------------------------------------------------------
-
-ajInpath
-No dependencies
-
-ajc
-\--- org.aspectj:aspectjtools:1.9.0.BETA-7
-
-apiElements - API elements for main. (n)
-No dependencies
-
-archives - Configuration for archive artifacts.
-No dependencies
-
-aspects
-\--- project :spring-orm
-     +--- project :spring-beans
-     |    \--- project :spring-core
-     |         \--- project :spring-jcl
-     +--- project :spring-core (*)
-     +--- project :spring-jdbc
-     |    +--- project :spring-beans (*)
-     |    +--- project :spring-core (*)
-     |    \--- project :spring-tx
-     |         +--- project :spring-beans (*)
-     |         \--- project :spring-core (*)
-     \--- project :spring-tx (*)
-
-compile - Dependencies for source set 'main' (deprecated, use 'implementation ' instead).
-\--- org.aspectj:aspectjweaver:1.8.11
-
-compileClasspath - Compile classpath for source set 'main'.
-+--- org.aspectj:aspectjweaver:1.8.11
-\--- com.google.code.findbugs:jsr305:3.0.2
-
-compileOnly - Compile only dependencies for source set 'main'.
-\--- com.google.code.findbugs:jsr305:3.0.2
-
-default - Configuration for default artifacts.
-\--- org.aspectj:aspectjweaver:1.8.11
-
-implementation - Implementation only dependencies for source set 'main'. (n)
-No dependencies
-
-kapt
-\--- org.jetbrains.kotlin:kotlin-annotation-processing:1.1.51
-     \--- org.jetbrains.kotlin:kotlin-stdlib:1.1.51
-          \--- org.jetbrains:annotations:13.0
-
-kaptTest
-\--- org.jetbrains.kotlin:kotlin-annotation-processing:1.1.51
-     \--- org.jetbrains.kotlin:kotlin-stdlib:1.1.51
-          \--- org.jetbrains:annotations:13.0
-
-optional
-+--- project :spring-aop
-|    +--- project :spring-beans
-|    |    \--- project :spring-core
-|    |         \--- project :spring-jcl
-|    \--- project :spring-core (*)
-+--- project :spring-beans (*)
-+--- project :spring-context
-|    +--- project :spring-aop (*)
-|    +--- project :spring-beans (*)
-|    +--- project :spring-core (*)
-|    \--- project :spring-expression
-|         \--- project :spring-core (*)
-+--- project :spring-context-support
-|    +--- project :spring-beans (*)
-|    +--- project :spring-context (*)
-|    \--- project :spring-core (*)
-+--- project :spring-orm
-|    +--- project :spring-beans (*)
-|    +--- project :spring-core (*)
-|    +--- project :spring-jdbc
-|    |    +--- project :spring-beans (*)
-|    |    +--- project :spring-core (*)
-|    |    \--- project :spring-tx
-|    |         +--- project :spring-beans (*)
-|    |         \--- project :spring-core (*)
-|    \--- project :spring-tx (*)
-+--- project :spring-tx (*)
-+--- javax.cache:cache-api:1.0.0
-\--- javax.transaction:javax.transaction-api:1.2
-
-provided
-No dependencies
-
-rt
-\--- org.aspectj:aspectjrt:1.9.0.BETA-7
-
-runtime - Runtime dependencies for source set 'main' (deprecated, use 'runtimeOnly ' instead).
-\--- org.aspectj:aspectjweaver:1.8.11
-
-runtimeClasspath - Runtime classpath of source set 'main'.
-\--- org.aspectj:aspectjweaver:1.8.11
-
-runtimeElements - Elements of runtime for main. (n)
-No dependencies
-
-runtimeOnly - Runtime only dependencies for source set 'main'. (n)
-No dependencies
-
-testCompile - Dependencies for source set 'test' (deprecated, use 'testImplementation ' instead).
-+--- org.aspectj:aspectjweaver:1.8.11
-+--- junit:junit:4.12
-+--- org.mockito:mockito-core:2.6.1
-|    +--- net.bytebuddy:byte-buddy:1.6.2
-|    +--- net.bytebuddy:byte-buddy-agent:1.6.2
-|    \--- org.objenesis:objenesis:2.4
-+--- com.nhaarman:mockito-kotlin:1.5.0
-+--- org.hamcrest:hamcrest-all:1.3
-+--- org.xmlunit:xmlunit-matchers:2.3.0
-|    +--- org.xmlunit:xmlunit-core:2.3.0
-|    \--- org.hamcrest:hamcrest-core:1.3
-+--- project :spring-core
-|    \--- project :spring-jcl
-+--- project :spring-test
-|    \--- project :spring-core (*)
-\--- javax.mail:javax.mail-api:1.6.0
-
-testCompileClasspath - Compile classpath for source set 'test'.
-+--- org.aspectj:aspectjweaver:1.8.11
-+--- junit:junit:4.12
-+--- org.mockito:mockito-core:2.6.1
-|    +--- net.bytebuddy:byte-buddy:1.6.2
-|    +--- net.bytebuddy:byte-buddy-agent:1.6.2
-|    \--- org.objenesis:objenesis:2.4
-+--- com.nhaarman:mockito-kotlin:1.5.0
-+--- org.hamcrest:hamcrest-all:1.3
-+--- org.xmlunit:xmlunit-matchers:2.3.0
-|    +--- org.xmlunit:xmlunit-core:2.3.0
-|    \--- org.hamcrest:hamcrest-core:1.3
-+--- project :spring-core
-|    \--- project :spring-jcl
-+--- project :spring-test
-|    \--- project :spring-core (*)
-+--- javax.mail:javax.mail-api:1.6.0
-\--- com.google.code.findbugs:jsr305:3.0.2
-
-testCompileOnly - Compile only dependencies for source set 'test'.
-\--- com.google.code.findbugs:jsr305:3.0.2
-
-testImplementation - Implementation only dependencies for source set 'test'. (n)
-No dependencies
-
-testRuntime - Runtime dependencies for source set 'test' (deprecated, use 'testRuntimeOnly ' instead).
-+--- org.aspectj:aspectjweaver:1.8.11
-+--- junit:junit:4.12
-+--- org.mockito:mockito-core:2.6.1
-|    +--- net.bytebuddy:byte-buddy:1.6.2
-|    +--- net.bytebuddy:byte-buddy-agent:1.6.2
-|    \--- org.objenesis:objenesis:2.4
-+--- com.nhaarman:mockito-kotlin:1.5.0
-+--- org.hamcrest:hamcrest-all:1.3
-+--- org.xmlunit:xmlunit-matchers:2.3.0
-|    +--- org.xmlunit:xmlunit-core:2.3.0
-|    \--- org.hamcrest:hamcrest-core:1.3
-+--- project :spring-core
-|    \--- project :spring-jcl
-+--- project :spring-test
-|    \--- project :spring-core (*)
-+--- javax.mail:javax.mail-api:1.6.0
-\--- org.apache.logging.log4j:log4j-core:2.9.1
-     \--- org.apache.logging.log4j:log4j-api:2.9.1
-
-testRuntimeClasspath - Runtime classpath of source set 'test'.
-+--- org.aspectj:aspectjweaver:1.8.11
-+--- junit:junit:4.12
-+--- org.mockito:mockito-core:2.6.1
-|    +--- net.bytebuddy:byte-buddy:1.6.2
-|    +--- net.bytebuddy:byte-buddy-agent:1.6.2
-|    \--- org.objenesis:objenesis:2.4
-+--- com.nhaarman:mockito-kotlin:1.5.0
-+--- org.hamcrest:hamcrest-all:1.3
-+--- org.xmlunit:xmlunit-matchers:2.3.0
-|    +--- org.xmlunit:xmlunit-core:2.3.0
-|    \--- org.hamcrest:hamcrest-core:1.3
-+--- project :spring-core
-|    \--- project :spring-jcl
-+--- project :spring-test
-|    \--- project :spring-core (*)
-+--- javax.mail:javax.mail-api:1.6.0
-\--- org.apache.logging.log4j:log4j-core:2.9.1
-     \--- org.apache.logging.log4j:log4j-api:2.9.1
-
-testRuntimeOnly - Runtime only dependencies for source set 'test'. (n)
-No dependencies
-
-(*) - dependencies omitted (listed previously)
-
-DETECT META DATA START
-rootProjectPath:/Users/ekerwin/Documents/source/spring-framework
-rootProjectGroup:org.springframework
-rootProjectName:spring
-rootProjectVersion:5.0.1.BUILD-SNAPSHOT
-projectPath:/Users/ekerwin/Documents/source/spring-framework/spring-aspects
-projectGroup:org.springframework
-projectName:spring-aspects
-projectVersion:5.0.1.BUILD-SNAPSHOT
-DETECT META DATA END
