--- detectable/src/test/resources/detectables/functional/gradle/spring-framework/spring_core_dependencyGraph.txt
@@ -1,189 +1,0 @@
-
-------------------------------------------------------------
-Project :spring-core - Spring Core
-------------------------------------------------------------
-
-apiElements - API elements for main. (n)
-No dependencies
-
-archives - Configuration for archive artifacts.
-No dependencies
-
-cglib
-\--- cglib:cglib:3.2.5
-
-compile - Dependencies for source set 'main' (deprecated, use 'implementation ' instead).
-\--- project :spring-jcl
-
-compileClasspath - Compile classpath for source set 'main'.
-+--- project :spring-jcl
-\--- com.google.code.findbugs:jsr305:3.0.2
-
-compileOnly - Compile only dependencies for source set 'main'.
-\--- com.google.code.findbugs:jsr305:3.0.2
-
-default - Configuration for default artifacts.
-\--- project :spring-jcl
-
-implementation - Implementation only dependencies for source set 'main'. (n)
-No dependencies
-
-jarjar
-\--- com.googlecode.jarjar:jarjar:1.3
-
-kapt
-\--- org.jetbrains.kotlin:kotlin-annotation-processing:1.1.51
-     \--- org.jetbrains.kotlin:kotlin-stdlib:1.1.51
-          \--- org.jetbrains:annotations:13.0
-
-kaptTest
-\--- org.jetbrains.kotlin:kotlin-annotation-processing:1.1.51
-     \--- org.jetbrains.kotlin:kotlin-stdlib:1.1.51
-          \--- org.jetbrains:annotations:13.0
-
-objenesis
-\--- org.objenesis:objenesis:2.6
-
-optional
-+--- net.sf.jopt-simple:jopt-simple:5.0.4
-+--- org.aspectj:aspectjweaver:1.8.11
-+--- org.jetbrains.kotlin:kotlin-reflect:1.1.51
-|    \--- org.jetbrains.kotlin:kotlin-stdlib:1.1.51
-|         \--- org.jetbrains:annotations:13.0
-+--- org.jetbrains.kotlin:kotlin-stdlib:1.1.51 (*)
-+--- io.projectreactor:reactor-core: -> 3.1.1.BUILD-SNAPSHOT
-|    \--- org.reactivestreams:reactive-streams:1.0.1
-+--- io.reactivex:rxjava:1.3.2
-+--- io.reactivex:rxjava-reactive-streams:1.2.1
-|    +--- io.reactivex:rxjava:1.2.2 -> 1.3.2
-|    \--- org.reactivestreams:reactive-streams:1.0.0 -> 1.0.1
-+--- io.reactivex.rxjava2:rxjava:2.1.5
-|    \--- org.reactivestreams:reactive-streams:1.0.1
-\--- io.netty:netty-buffer: -> 4.1.16.Final
-     \--- io.netty:netty-common:4.1.16.Final
-
-provided
-No dependencies
-
-runtime - Runtime dependencies for source set 'main' (deprecated, use 'runtimeOnly ' instead).
-\--- project :spring-jcl
-
-runtimeClasspath - Runtime classpath of source set 'main'.
-\--- project :spring-jcl
-
-runtimeElements - Elements of runtime for main. (n)
-No dependencies
-
-runtimeOnly - Runtime only dependencies for source set 'main'. (n)
-No dependencies
-
-testCompile - Dependencies for source set 'test' (deprecated, use 'testImplementation ' instead).
-+--- project :spring-jcl
-+--- junit:junit:4.12
-+--- org.mockito:mockito-core:2.6.1
-|    +--- net.bytebuddy:byte-buddy:1.6.2
-|    +--- net.bytebuddy:byte-buddy-agent:1.6.2
-|    \--- org.objenesis:objenesis:2.4
-+--- com.nhaarman:mockito-kotlin:1.5.0
-+--- org.hamcrest:hamcrest-all:1.3
-+--- org.xmlunit:xmlunit-matchers:2.3.0
-|    +--- org.xmlunit:xmlunit-core:2.3.0
-|    \--- org.hamcrest:hamcrest-core:1.3
-+--- io.projectreactor:reactor-test: -> 3.1.1.BUILD-SNAPSHOT
-|    \--- io.projectreactor:reactor-core:3.1.1.BUILD-SNAPSHOT
-|         \--- org.reactivestreams:reactive-streams:1.0.1
-+--- javax.xml.bind:jaxb-api:2.3.0
-+--- org.apache.tomcat.embed:tomcat-embed-core:8.5.23
-|    \--- org.apache.tomcat:tomcat-annotations-api:8.5.23
-\--- com.fasterxml.woodstox:woodstox-core:5.0.3
-     \--- org.codehaus.woodstox:stax2-api:3.1.4
-
-testCompileClasspath - Compile classpath for source set 'test'.
-+--- project :spring-jcl
-+--- junit:junit:4.12
-+--- org.mockito:mockito-core:2.6.1
-|    +--- net.bytebuddy:byte-buddy:1.6.2
-|    +--- net.bytebuddy:byte-buddy-agent:1.6.2
-|    \--- org.objenesis:objenesis:2.4
-+--- com.nhaarman:mockito-kotlin:1.5.0
-+--- org.hamcrest:hamcrest-all:1.3
-+--- org.xmlunit:xmlunit-matchers:2.3.0
-|    +--- org.xmlunit:xmlunit-core:2.3.0
-|    \--- org.hamcrest:hamcrest-core:1.3
-+--- io.projectreactor:reactor-test: -> 3.1.1.BUILD-SNAPSHOT
-|    \--- io.projectreactor:reactor-core:3.1.1.BUILD-SNAPSHOT
-|         \--- org.reactivestreams:reactive-streams:1.0.1
-+--- javax.xml.bind:jaxb-api:2.3.0
-+--- org.apache.tomcat.embed:tomcat-embed-core:8.5.23
-|    \--- org.apache.tomcat:tomcat-annotations-api:8.5.23
-+--- com.fasterxml.woodstox:woodstox-core:5.0.3
-|    \--- org.codehaus.woodstox:stax2-api:3.1.4
-\--- com.google.code.findbugs:jsr305:3.0.2
-
-testCompileOnly - Compile only dependencies for source set 'test'.
-\--- com.google.code.findbugs:jsr305:3.0.2
-
-testImplementation - Implementation only dependencies for source set 'test'. (n)
-No dependencies
-
-testRuntime - Runtime dependencies for source set 'test' (deprecated, use 'testRuntimeOnly ' instead).
-+--- project :spring-jcl
-+--- junit:junit:4.12
-+--- org.mockito:mockito-core:2.6.1
-|    +--- net.bytebuddy:byte-buddy:1.6.2
-|    +--- net.bytebuddy:byte-buddy-agent:1.6.2
-|    \--- org.objenesis:objenesis:2.4
-+--- com.nhaarman:mockito-kotlin:1.5.0
-+--- org.hamcrest:hamcrest-all:1.3
-+--- org.xmlunit:xmlunit-matchers:2.3.0
-|    +--- org.xmlunit:xmlunit-core:2.3.0
-|    \--- org.hamcrest:hamcrest-core:1.3
-+--- io.projectreactor:reactor-test: -> 3.1.1.BUILD-SNAPSHOT
-|    \--- io.projectreactor:reactor-core:3.1.1.BUILD-SNAPSHOT
-|         \--- org.reactivestreams:reactive-streams:1.0.1
-+--- javax.xml.bind:jaxb-api:2.3.0
-+--- org.apache.tomcat.embed:tomcat-embed-core:8.5.23
-|    \--- org.apache.tomcat:tomcat-annotations-api:8.5.23
-+--- com.fasterxml.woodstox:woodstox-core:5.0.3
-|    \--- org.codehaus.woodstox:stax2-api:3.1.4
-\--- org.apache.logging.log4j:log4j-core:2.9.1
-     \--- org.apache.logging.log4j:log4j-api:2.9.1
-
-testRuntimeClasspath - Runtime classpath of source set 'test'.
-+--- project :spring-jcl
-+--- junit:junit:4.12
-+--- org.mockito:mockito-core:2.6.1
-|    +--- net.bytebuddy:byte-buddy:1.6.2
-|    +--- net.bytebuddy:byte-buddy-agent:1.6.2
-|    \--- org.objenesis:objenesis:2.4
-+--- com.nhaarman:mockito-kotlin:1.5.0
-+--- org.hamcrest:hamcrest-all:1.3
-+--- org.xmlunit:xmlunit-matchers:2.3.0
-|    +--- org.xmlunit:xmlunit-core:2.3.0
-|    \--- org.hamcrest:hamcrest-core:1.3
-+--- io.projectreactor:reactor-test: -> 3.1.1.BUILD-SNAPSHOT
-|    \--- io.projectreactor:reactor-core:3.1.1.BUILD-SNAPSHOT
-|         \--- org.reactivestreams:reactive-streams:1.0.1
-+--- javax.xml.bind:jaxb-api:2.3.0
-+--- org.apache.tomcat.embed:tomcat-embed-core:8.5.23
-|    \--- org.apache.tomcat:tomcat-annotations-api:8.5.23
-+--- com.fasterxml.woodstox:woodstox-core:5.0.3
-|    \--- org.codehaus.woodstox:stax2-api:3.1.4
-\--- org.apache.logging.log4j:log4j-core:2.9.1
-     \--- org.apache.logging.log4j:log4j-api:2.9.1
-
-testRuntimeOnly - Runtime only dependencies for source set 'test'. (n)
-No dependencies
-
-(*) - dependencies omitted (listed previously)
-
-DETECT META DATA START
-rootProjectPath:/Users/ekerwin/Documents/source/spring-framework
-rootProjectGroup:org.springframework
-rootProjectName:spring
-rootProjectVersion:5.0.1.BUILD-SNAPSHOT
-projectPath:/Users/ekerwin/Documents/source/spring-framework/spring-core
-projectGroup:org.springframework
-projectName:spring-core
-projectVersion:5.0.1.BUILD-SNAPSHOT
-DETECT META DATA END
