+++ detectable/src/test/resources/detectables/functional/gradle/spring_aop_dependencyGraph.txt
@@ -1,0 +1,164 @@
+
+------------------------------------------------------------
+Project :spring-aop - Spring AOP
+------------------------------------------------------------
+
+apiElements - API elements for main. (n)
+No dependencies
+
+archives - Configuration for archive artifacts.
+No dependencies
+
+compile - Dependencies for source set 'main' (deprecated, use 'implementation ' instead).
++--- project :spring-beans
+|    \--- project :spring-core
+|         \--- project :spring-jcl
+\--- project :spring-core (*)
+
+compileClasspath - Compile classpath for source set 'main'.
++--- project :spring-beans
+|    \--- project :spring-core
+|         \--- project :spring-jcl
++--- project :spring-core (*)
+\--- com.google.code.findbugs:jsr305:3.0.2
+
+compileOnly - Compile only dependencies for source set 'main'.
+\--- com.google.code.findbugs:jsr305:3.0.2
+
+default - Configuration for default artifacts.
++--- project :spring-beans
+|    \--- project :spring-core
+|         \--- project :spring-jcl
+\--- project :spring-core (*)
+
+implementation - Implementation only dependencies for source set 'main'. (n)
+No dependencies
+
+kapt
+\--- org.jetbrains.kotlin:kotlin-annotation-processing:1.1.51
+     \--- org.jetbrains.kotlin:kotlin-stdlib:1.1.51
+          \--- org.jetbrains:annotations:13.0
+
+kaptTest
+\--- org.jetbrains.kotlin:kotlin-annotation-processing:1.1.51
+     \--- org.jetbrains.kotlin:kotlin-stdlib:1.1.51
+          \--- org.jetbrains:annotations:13.0
+
+optional
++--- org.aspectj:aspectjweaver:1.8.11
++--- org.apache.commons:commons-pool2:2.4.2
+\--- com.jamonapi:jamon:2.81
+     \--- com.hazelcast:hazelcast-all:3.2.5
+          \--- net.sourceforge.findbugs:annotations:1.3.2
+
+provided
+No dependencies
+
+runtime - Runtime dependencies for source set 'main' (deprecated, use 'runtimeOnly ' instead).
++--- project :spring-beans
+|    \--- project :spring-core
+|         \--- project :spring-jcl
+\--- project :spring-core (*)
+
+runtimeClasspath - Runtime classpath of source set 'main'.
++--- project :spring-beans
+|    \--- project :spring-core
+|         \--- project :spring-jcl
+\--- project :spring-core (*)
+
+runtimeElements - Elements of runtime for main. (n)
+No dependencies
+
+runtimeOnly - Runtime only dependencies for source set 'main'. (n)
+No dependencies
+
+testCompile - Dependencies for source set 'test' (deprecated, use 'testImplementation ' instead).
++--- project :spring-beans
+|    \--- project :spring-core
+|         \--- project :spring-jcl
++--- project :spring-core (*)
++--- junit:junit:4.12
++--- org.mockito:mockito-core:2.6.1
+|    +--- net.bytebuddy:byte-buddy:1.6.2
+|    +--- net.bytebuddy:byte-buddy-agent:1.6.2
+|    \--- org.objenesis:objenesis:2.4
++--- com.nhaarman:mockito-kotlin:1.5.0
++--- org.hamcrest:hamcrest-all:1.3
+\--- org.xmlunit:xmlunit-matchers:2.3.0
+     +--- org.xmlunit:xmlunit-core:2.3.0
+     \--- org.hamcrest:hamcrest-core:1.3
+
+testCompileClasspath - Compile classpath for source set 'test'.
++--- project :spring-beans
+|    \--- project :spring-core
+|         \--- project :spring-jcl
++--- project :spring-core (*)
++--- junit:junit:4.12
++--- org.mockito:mockito-core:2.6.1
+|    +--- net.bytebuddy:byte-buddy:1.6.2
+|    +--- net.bytebuddy:byte-buddy-agent:1.6.2
+|    \--- org.objenesis:objenesis:2.4
++--- com.nhaarman:mockito-kotlin:1.5.0
++--- org.hamcrest:hamcrest-all:1.3
++--- org.xmlunit:xmlunit-matchers:2.3.0
+|    +--- org.xmlunit:xmlunit-core:2.3.0
+|    \--- org.hamcrest:hamcrest-core:1.3
+\--- com.google.code.findbugs:jsr305:3.0.2
+
+testCompileOnly - Compile only dependencies for source set 'test'.
+\--- com.google.code.findbugs:jsr305:3.0.2
+
+testImplementation - Implementation only dependencies for source set 'test'. (n)
+No dependencies
+
+testRuntime - Runtime dependencies for source set 'test' (deprecated, use 'testRuntimeOnly ' instead).
++--- project :spring-beans
+|    \--- project :spring-core
+|         \--- project :spring-jcl
++--- project :spring-core (*)
++--- junit:junit:4.12
++--- org.mockito:mockito-core:2.6.1
+|    +--- net.bytebuddy:byte-buddy:1.6.2
+|    +--- net.bytebuddy:byte-buddy-agent:1.6.2
+|    \--- org.objenesis:objenesis:2.4
++--- com.nhaarman:mockito-kotlin:1.5.0
++--- org.hamcrest:hamcrest-all:1.3
++--- org.xmlunit:xmlunit-matchers:2.3.0
+|    +--- org.xmlunit:xmlunit-core:2.3.0
+|    \--- org.hamcrest:hamcrest-core:1.3
+\--- org.apache.logging.log4j:log4j-core:2.9.1
+     \--- org.apache.logging.log4j:log4j-api:2.9.1
+
+testRuntimeClasspath - Runtime classpath of source set 'test'.
++--- project :spring-beans
+|    \--- project :spring-core
+|         \--- project :spring-jcl
++--- project :spring-core (*)
++--- junit:junit:4.12
++--- org.mockito:mockito-core:2.6.1
+|    +--- net.bytebuddy:byte-buddy:1.6.2
+|    +--- net.bytebuddy:byte-buddy-agent:1.6.2
+|    \--- org.objenesis:objenesis:2.4
++--- com.nhaarman:mockito-kotlin:1.5.0
++--- org.hamcrest:hamcrest-all:1.3
++--- org.xmlunit:xmlunit-matchers:2.3.0
+|    +--- org.xmlunit:xmlunit-core:2.3.0
+|    \--- org.hamcrest:hamcrest-core:1.3
+\--- org.apache.logging.log4j:log4j-core:2.9.1
+     \--- org.apache.logging.log4j:log4j-api:2.9.1
+
+testRuntimeOnly - Runtime only dependencies for source set 'test'. (n)
+No dependencies
+
+(*) - dependencies omitted (listed previously)
+
+DETECT META DATA START
+rootProjectPath:/Users/ekerwin/Documents/source/spring-framework
+rootProjectGroup:org.springframework
+rootProjectName:spring
+rootProjectVersion:5.0.1.BUILD-SNAPSHOT
+projectPath:/Users/ekerwin/Documents/source/spring-framework/spring-aop
+projectGroup:org.springframework
+projectName:spring-aop
+projectVersion:5.0.1.BUILD-SNAPSHOT
+DETECT META DATA END
